Refactor types to new source file
